// SDKs/googleapis.com/language/v1beta2/typescript/src/sdk/models/shared/analyzesentimentrequest.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import { Document } from "./document";


export enum AnalyzeSentimentRequestEncodingTypeEnum {
    None = "NONE",
    Utf8 = "UTF8",
    Utf16 = "UTF16",
    Utf32 = "UTF32"
}


// AnalyzeSentimentRequest
/** 
 * The sentiment analysis request message.
**/
export class AnalyzeSentimentRequest extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=document" })
  document?: Document;

  @SpeakeasyMetadata({ data: "json, name=encodingType" })
  encodingType?: AnalyzeSentimentRequestEncodingTypeEnum;
}