// SDKs/googleapis.com/language/v1beta2/typescript/src/sdk/models/operations/languagedocumentsanalyzesentiment.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import * as shared from "../shared";


export class LanguageDocumentsAnalyzeSentimentQueryParams extends SpeakeasyBase {
  @Metadata({ data: "queryParam, style=form;explode=true;name=$.xgafv" })
  dollarXgafv?: shared.XgafvEnum;

  @Metadata({ data: "queryParam, style=form;explode=true;name=access_token" })
  accessToken?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=alt" })
  alt?: shared.AltEnum;

  @Metadata({ data: "queryParam, style=form;explode=true;name=callback" })
  callback?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=fields" })
  fields?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=key" })
  key?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=oauth_token" })
  oauthToken?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=prettyPrint" })
  prettyPrint?: boolean;

  @Metadata({ data: "queryParam, style=form;explode=true;name=quotaUser" })
  quotaUser?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=uploadType" })
  uploadType?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=upload_protocol" })
  uploadProtocol?: string;
}


export class LanguageDocumentsAnalyzeSentimentSecurityOption1 extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=oauth2" })
  oauth2: shared.SchemeOauth2;

  @Metadata({ data: "security, scheme=true;type=oauth2" })
  oauth2c: shared.SchemeOauth2c;
}


export class LanguageDocumentsAnalyzeSentimentSecurityOption2 extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=oauth2" })
  oauth2: shared.SchemeOauth2;

  @Metadata({ data: "security, scheme=true;type=oauth2" })
  oauth2c: shared.SchemeOauth2c;
}


export class LanguageDocumentsAnalyzeSentimentSecurity extends SpeakeasyBase {
  @Metadata({ data: "security, option=true" })
  option1?: LanguageDocumentsAnalyzeSentimentSecurityOption1;

  @Metadata({ data: "security, option=true" })
  option2?: LanguageDocumentsAnalyzeSentimentSecurityOption2;
}


export class LanguageDocumentsAnalyzeSentimentRequest extends SpeakeasyBase {
  @Metadata()
  queryParams: LanguageDocumentsAnalyzeSentimentQueryParams;

  @Metadata({ data: "request, media_type=application/json" })
  request?: shared.AnalyzeSentimentRequest;

  @Metadata()
  security: LanguageDocumentsAnalyzeSentimentSecurity;
}


export class LanguageDocumentsAnalyzeSentimentResponse extends SpeakeasyBase {
  @Metadata()
  analyzeSentimentResponse?: shared.AnalyzeSentimentResponse;

  @Metadata()
  contentType: string;

  @Metadata()
  statusCode: number;
}
